[junu0516/airbnb#31] Ref: 숙소찾기결과 ViewController 에서 하위View 를 파일로 분리
