+++ iOS/Airbnb/RoomList/RoomListViewController.swift
@@ -5,6 +5,22 @@ class RoomListViewController: UIViewController {
 
     private var useCase: RoomListUseCase?
     
+    convenience init(useCase: RoomListUseCase) {
+        self.init()
+        self.useCase = useCase
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        self.title = "숙소 찾기"
+        self.view.backgroundColor = .white
+        setupViews()
+    }
+    
+    // MARK: - Views
+    private let headerView = SearchResultRoomsHeaderView()
+    
     private lazy var collectionView: UICollectionView = {
         let layout = UICollectionViewFlowLayout()
         layout.sectionInset = UIEdgeInsets(top: 20, left: 0, bottom: 0, right: 0)
@@ -33,21 +49,7 @@ class RoomListViewController: UIViewController {
         return button
     }()
     
-    private let headerView = SearchResultRoomsHeaderView()
-    
-    convenience init(useCase: RoomListUseCase) {
-        self.init()
-        self.useCase = useCase
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        self.title = "숙소 찾기"
-        self.view.backgroundColor = .white
-        setupViews()
-    }
-    
+    // MARK: - private functions
     private func setupViews() {
         view.addSubview(headerView)
         headerView.translatesAutoresizingMaskIntoConstraints = false
@@ -98,82 +100,4 @@ extension RoomListViewController: UICollectionViewDelegate {
         viewController.modalPresentationStyle = .fullScreen
         self.present(viewController, animated: false)
     }
-}
-
-
-// MARK: - SearchResultRoomCell
-class SearchResultRoomCell: UICollectionViewCell {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupViews()
-        self.backgroundColor = .green
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func updateViews(title: String) {
-        self.titleLabel.text = title
-    }
-    
-    private let titleLabel: UILabel = {
-        let label = UILabel()
-        label.text = "Title"
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.backgroundColor = .lightGray
-        return label
-    }()
-    
-    private func setupViews() {
-        self.addSubview(titleLabel)
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: topAnchor),
-            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-            titleLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
-            titleLabel.bottomAnchor.constraint(equalTo: bottomAnchor)
-        ])
-    }
-}
-
-
-// MARK: - SearchResultRoomsHeaderView
-class SearchResultRoomsHeaderView: UIView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupViews()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setupViews() {
-        let stackView = UIStackView(arrangedSubviews: [detailDescriptionLabel, roomCountLabel])
-        stackView.axis = .vertical
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        
-        self.addSubview(stackView)
-        NSLayoutConstraint.activate([
-            stackView.topAnchor.constraint(equalTo: topAnchor),
-            stackView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            stackView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            self.bottomAnchor.constraint(equalTo: stackView.bottomAnchor)
-        ])
-    }
-    
-    private let detailDescriptionLabel: UILabel = {
-        let label = UILabel()
-        label.text = "5월 25일 ~ 5월 28일 ・ 게스트 3명"
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    
-    private let roomCountLabel: UILabel = {
-        let label = UILabel()
-        label.text = "300개 이상의 숙소"
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
 }
