+++ iOS/Airbnb/RoomList/View/SearchResultRoomCell.swift
@@ -1,0 +1,35 @@
+import UIKit
+
+final class SearchResultRoomCell: UICollectionViewCell {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupViews()
+        self.backgroundColor = .green
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func updateViews(title: String) {
+        self.titleLabel.text = title
+    }
+    
+    private let titleLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Title"
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.backgroundColor = .lightGray
+        return label
+    }()
+    
+    private func setupViews() {
+        self.addSubview(titleLabel)
+        NSLayoutConstraint.activate([
+            titleLabel.topAnchor.constraint(equalTo: topAnchor),
+            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
+            titleLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
+            titleLabel.bottomAnchor.constraint(equalTo: bottomAnchor)
+        ])
+    }
+}
